Refactored code and claned up some functions
