+++ Packages/com.nickmaltbie.openkcc/Editor/HumanoidFootIKEditor.cs
@@ -6,7 +6,6 @@ using nickmaltbie.OpenKCC.Animation;
 using Unity.EditorCoroutines.Editor;
 using UnityEditor;
 using UnityEngine;
-using UnityEngine.UIElements;
 
 namespace nickmaltbie.OpenKCC.Editor
 {
@@ -75,14 +74,14 @@ namespace nickmaltbie.OpenKCC.Editor
 
                 ClipAnimationInfoCurve leftInfoCurve = new ClipAnimationInfoCurve();
                 ClipAnimationInfoCurve rightInfoCurve = new ClipAnimationInfoCurve();
-                leftInfoCurve.name = HumanoidFootIK.LeftFootIKWeight;
-                rightInfoCurve.name = HumanoidFootIK.RightFootIKWeight;
+                leftInfoCurve.name = FootTarget.LeftFootIKWeight;
+                rightInfoCurve.name = FootTarget.RightFootIKWeight;
 
                 var importer = AssetImporter.GetAtPath(assetPath) as ModelImporter;
                 ModelImporterClipAnimation[] anims = importer.clipAnimations;
                 int animIndex = Enumerable.Range(0, anims.Length).FirstOrDefault(i => anims[i].name == clip.name);
 
-                anims[animIndex].curves = anims[animIndex].curves.Where(c => c.name != HumanoidFootIK.LeftFootIKWeight && c.name != HumanoidFootIK.RightFootIKWeight).ToArray();
+                anims[animIndex].curves = anims[animIndex].curves.Where(c => c.name != FootTarget.LeftFootIKWeight && c.name != FootTarget.RightFootIKWeight).ToArray();
                 importer.clipAnimations = anims;
 
                 // So you actually need to cleanup the animations first...
@@ -158,7 +157,7 @@ namespace nickmaltbie.OpenKCC.Editor
                 rightInfoCurve.curve = new AnimationCurve(rightFootKeys);
 
                 anims[animIndex].curves = Enumerable.Concat(
-                    anims[animIndex].curves.Where(c => c.name != HumanoidFootIK.LeftFootIKWeight && c.name != HumanoidFootIK.RightFootIKWeight),
+                    anims[animIndex].curves.Where(c => c.name != FootTarget.LeftFootIKWeight && c.name != FootTarget.RightFootIKWeight),
                     new [] { leftInfoCurve, rightInfoCurve }).ToArray();
                 importer.clipAnimations = anims;
