+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Animation/Foot.cs
@@ -1,0 +1,9 @@
+
+namespace nickmaltbie.OpenKCC.Animation
+{
+    public enum Foot
+    {
+        LeftFoot,
+        RightFoot,
+    }
+}
