+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Animation/Foot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d7d53e8c2f52e04784eeb196086453c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
