+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Animation/FootState.cs
@@ -1,0 +1,9 @@
+
+namespace nickmaltbie.OpenKCC.Animation
+{
+    public enum FootState
+    {
+        Released,
+        Grounded
+    }
+}
