+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Animation/FootState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 14488c8a3ea1761479f6151b3b76e385
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
