+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Animation/FootTarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3423310ce0883834ab7d77d971b992f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
